[TIMOB-25238] Fix int to byte conversion

+++ android/titanium/src/java/org/appcelerator/titanium/view/TiUIView.java
@@ -884,7 +884,7 @@ public abstract class TiUIView
 
 					// TIMOB-24898: disable HW acceleration to allow transparency
 					// when the backgroundColor alpha channel has been set
-					byte bgAlpha = bgColor != null ? (byte)(bgColor >> 24) : 0xFF;
+					byte bgAlpha = bgColor != null ? (byte)(bgColor >> 24) : (byte)0xFF;
 					if (bgAlpha != 0xFF) {
 						disableHWAcceleration();
 					}
@@ -1439,7 +1439,7 @@ public abstract class TiUIView
 
 			// TIMOB-24898: disable HW acceleration to allow transparency
 			// when the backgroundColor alpha channel has been set
-			byte bgAlpha = bgColor != null ? (byte)(bgColor >> 24) : 0xFF;
+			byte bgAlpha = bgColor != null ? (byte)(bgColor >> 24) : (byte)0xFF;
 			if (bgAlpha != 0xFF) {
 				disableHWAcceleration();
 			}
